Moves views
