+++ lib/assets/javascripts/cartodb3/deep-insights-integrations.js
@@ -2,7 +2,7 @@ var _ = require('underscore');
 var linkLayerInfowindow = require('./deep-insights-integration/link-layer-infowindow');
 var linkLayerTooltip = require('./deep-insights-integration/link-layer-tooltip');
 var LegendManager = require('./deep-insights-integration/legend-manager');
-var AnalysisNotifications = require('./analysis-notifications');
+var AnalysisNotifications = require('./editor/layers/analysis-views/analysis-notifications');
 var WidgetsNotifications = require('./widgets-notifications');
 var AnalysisOnboardingLauncher = require('./components/onboardings/analysis/analysis-launcher');
 var VisNotifications = require('./vis-notifications');
