+++ lib/assets/javascripts/cartodb3/editor/layers/analysis-views/analysis-notifications.js
@@ -1,5 +1,5 @@
 var _ = require('underscore');
-var Notifier = require('./components/notifier/notifier');
+var Notifier = require('../../../components/notifier/notifier');
 var AnalysisNotificationErrorMessageHandler = require('./analysis-notification-error-message-handler');
 
 var DEFAULT_DELAY = Notifier.DEFAULT_DELAY;
