+++ lib/assets/test/jasmine/cartodb3/analysis-notifications.spec.js
@@ -1,7 +1,7 @@
 var _ = require('underscore');
 var Backbone = require('backbone');
 var Notifier = require('../../../javascripts/cartodb3/components/notifier/notifier');
-var AnalysisNotifications = require('../../../javascripts/cartodb3/analysis-notifications');
+var AnalysisNotifications = require('../../../javascripts/cartodb3/editor/layers/analysis-views/analysis-notifications');
 
 describe('AnalysisNotifications.track', function () {
   beforeEach(function () {
