+++ lib/assets/test/spec/cartodb3/editor/layers/analysis-view/analysis-notification-error-message-handler.spec.js
@@ -1,5 +1,5 @@
 var cdb = require('cartodb.js');
-var AnalysisNotificationErrorMessageHandler = require('../../../../../../javascripts/cartodb3/analysis-notification-error-message-handler');
+var AnalysisNotificationErrorMessageHandler = require('../../../../../../javascripts/cartodb3/editor/layers/analysis-views/analysis-notification-error-message-handler');
 
 describe('AnalysisNotificationErrorMessageHandler.extractErrorMessage', function () {
   it('should extract the right error message', function () {
@@ -10,6 +10,10 @@ describe('AnalysisNotificationErrorMessageHandler.extractErrorMessage', function
       }
     });
     expect(AnalysisNotificationErrorMessageHandler.extractErrorMessage(analysisDefinitionNodeModel)).toEqual('notifications.analysis.failed: There was an error trying to obtain route using mapzen provider: Timeout reading from socket');
+
+    analysisDefinitionNodeModel.set('error', { message: 'InternalError: cdb_dataservices_server._obs_getmeasure(7): [obs_2016_10_14_65e534f91e] REMOTE ERROR: Error performing intersection: TopologyException: found non-noded intersection between LINESTRING (-96.5796 32.577, -96.579 32.5768) and LINESTRING (-96.579 32.5768, -96.579 32.5768) at -96.578972198741511 32.576761407730309' });
+
+    expect(AnalysisNotificationErrorMessageHandler.extractErrorMessage(analysisDefinitionNodeModel)).toEqual('found non-noded intersection between LINESTRING (-96.5796 32.577, -96.579 32.5768) and LINESTRING (-96.579 32.5768, -96.579 32.5768) at -96.578972198741511 32.576761407730309');
 
     analysisDefinitionNodeModel.set('error', { message: 'There was an error' });
     expect(AnalysisNotificationErrorMessageHandler.extractErrorMessage(analysisDefinitionNodeModel)).toEqual('notifications.analysis.failed: There was an error');
